Hide scrollbars when content collapsed
I have noticed in chrome when testing for several mobile browsers that if there is overflow content when the sidebar is collapsed that scrollbar can remain visible for the active pane and it gets placed above the buttons.  This small change forces the scrollbar to hide when the sidebar is collapsed.