JS/leaflet: add function for adding/removing click listeners
this._toggleTabClick(tab) en-/disables the click listener for the given tab,
depending wether the sidebar is added to the map or not.
this._toggleCloseClick(closeButton) does the same for closebuttons.

The _onClick & _onCloseClick functions are now encapsulated in the functions above,
as there is no need anymore to use them directly