Integrated support for OL5 based on ES modules.
OpenLayers 5 is implemented as a set of ES modules, requiring
major changes to the sidebar. Depending on how OL5 is used, the
sidebar will either be used as an ES module as well, or, in case the
full build of OL (ol.js) is used, a version compiled via webpack for
direct usage in a browser environment is required.