Add installation and usage documentation
In order to guide new users as to how to use the package it's helpful to
have basic installation and usage documentation.  This patch adds such
documentation and is based upon the latest release (v0.4.0) which works
with OpenLayers 4.6.5 and hence the example code will "just work" for
the user.

This patch addresses the issues raised in #50 and #104.