Add basic tests of Sidebar with OpenLayers
These tests currently only work in OL3.  Unfortunately I couldn't get
the tests to work in OL4-6 yet but hope these can be added over time
(there are issues with the `canvas` element in `node` that I've not yet
worked out).  The openlayers module is explicitly left out of the
devDependencies because I think it's more likely that the continuous
integration environment will install the relevant OL version and then
run the test suite; hence it's not necessary to tie down developers with
a particular OL version as part of development.

I've chosen to use "jest" as the testing framework here since it seems
to be the most popular testing framework currently and it's used by
other OpenLayers-related projects such as ol-contextmenu.

Also note that it was necessary to load the OpenLayers module explicitly
within the sidebar-v2 module itself so that OpenLayers was available for
the code running the tests.  Since the loading code only runs if
node/jsdom are available, it doesn't have an influence on the main of
code which runs in the browser.